# api/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'OUAGA CHAP API Documentation'
  description: 'API de la plateforme OUAGA CHAP — livraison de repas à Ouagadougou. Gestion des commandes, paiements, suivi en temps réel et notifications.'
  version: 1.0.0
servers:
  -
    url: 'https://ouagachap.pro'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Authentification via Laravel Sanctum. Envoyez le header `Authorization: Bearer {token}` obtenu via `POST /api/v1/auth/login`.'
security:
  -
    default: []
paths:
...
  /api/v1/orders:
    post:
      summary: 'Créer une commande'
      operationId: crerUneCommande
      description: 'Crée une nouvelle commande de livraison. Réservé aux clients.'
      parameters: []
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  message: 'Commande créée avec succès.'
                  data:
                    id: uuid
                    order_number: OC-20260120-ABCD
                    status: pending
                    total_price: 1500
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Commande créée avec succès.'
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: uuid
                      order_number:
                        type: string
                        example: OC-20260120-ABCD
                      status:
                        type: string
                        example: pending
                      total_price:
                        type: integer
                        example: 1500
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'Seuls les clients peuvent créer des commandes.'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Seuls les clients peuvent créer des commandes.'
      tags:
        - Commandes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pickup_address:
                  type: string
                  description: 'validation.min validation.max.'
                  example: b
                pickup_latitude:
                  type: number
                  description: 'Must match the regex /^-?\d{1,2}\.\d{4,8}$/. validation.between.'
                  example: -89
                pickup_longitude:
                  type: number
                  description: 'Must match the regex /^-?\d{1,3}\.\d{4,8}$/. validation.between.'
                  example: -180
                pickup_contact_name:
                  type: string
                  description: 'validation.min validation.max.'
                  example: z
                pickup_contact_phone:
                  type: string
                  description: 'Must match the regex /^(\+226|00226)?[0-9]{8}$/.'
                  example: '+22642559314'
                pickup_instructions:
                  type: string
                  description: validation.max.
                  example: k
                  nullable: true
                dropoff_address:
                  type: string
                  description: 'validation.min validation.max.'
                  example: h
                dropoff_latitude:
                  type: number
                  description: 'Must match the regex /^-?\d{1,2}\.\d{4,8}$/. validation.between.'
                  example: -90
                dropoff_longitude:
                  type: number
                  description: 'Must match the regex /^-?\d{1,3}\.\d{4,8}$/. validation.between.'
                  example: -180
                dropoff_contact_name:
                  type: string
                  description: 'validation.min validation.max.'
                  example: 'y'
                dropoff_contact_phone:
                  type: string
                  description: 'Must match the regex /^(\+226|00226)?[0-9]{8}$/.'
                  example: '+22642559314'
                dropoff_instructions:
                  type: string
                  description: validation.max.
                  example: k
                  nullable: true
                package_description:
                  type: string
                  description: 'validation.min validation.max.'
                  example: h
                package_size:
                  type: string
                  description: ''
                  example: small
                  enum:
                    - small
                    - medium
                    - large
                zone_id:
                  type: integer
                  description: 'The <code>id</code> of an existing record in the zones table.'
                  example: 16
                  nullable: true
              required:
                - pickup_address
                - pickup_latitude
                - pickup_longitude
                - pickup_contact_name
                - pickup_contact_phone
                - dropoff_address
                - dropoff_latitude
                - dropoff_longitude
                - dropoff_contact_name
                - dropoff_contact_phone
                - package_description
    get:
      summary: 'Lister mes commandes'
      operationId: listerMesCommandes
      description: 'Retourne la liste paginée des commandes du client connecté.'
      parameters:
        -
          in: query
          name: status
          description: 'Filtrer par statut (pending, assigned, picked_up, delivered, cancelled).'
          example: pending
          required: false
          schema:
            type: string
            description: 'Filtrer par statut (pending, assigned, picked_up, delivered, cancelled).'
            example: pending
        -
          in: query
          name: per_page
          description: 'Nombre de résultats par page.'
          example: 15
          required: false
          schema:
            type: integer
            description: 'Nombre de résultats par page.'
            example: 15
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  message: 'Commandes récupérées.'
                  data: []
                  meta:
                    current_page: 1
                    last_page: 1
                    per_page: 15
                    total: 0
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Commandes récupérées.'
                  data:
                    type: array
                    example: []
                  meta:
                    type: object
                    properties:
                      current_page:
                        type: integer
                        example: 1
                      last_page:
                        type: integer
                        example: 1
                      per_page:
                        type: integer
                        example: 15
                      total:
                        type: integer
                        example: 0
      tags:
        - Commandes
...